Created new interfaces to untangle the large GuideOcelotServer protobuf interface. Moved a lot of code around so the functions would be file-local before attempting to split the package
