+++ router/admin/ocelot_server.go
@@ -1,0 +1,8 @@
+package admin
+
+type OcelotServer interface {
+	SecretInterface
+	RepoInterface
+	StatusInterface
+	BuildInterface
+}
